Added Type Checking to SymbolTable.hs
The display of the type checking is still very primitive. There are plans for improving it by including the Posn wrapper from Alex and Happy, which would pass on to the AST. Considering that the current display for the type checking errors shows the expression where it occurred, this change should not prove difficult.
Also, the Main.hs file has been changed in order to accommodate the changes in the result of the SymbolTable.hs: a fix for the picking of the results, and the creation of a 4th option for only presenting the type checking result. An empty list means no errors, while each present element means a problem of its own. For clarity, the expected type and what it mismatched with give some company to each element (Expression) of the error list, thus creating a triple tuple.