Big Changes to SymbolTable and some minor tweaks
The parsing (Parser.y) for Get_Line has been fixed. Now it's an ID for each argument of the function.
PrintAST.hs has been fixed in response to an error caused by the change in the data of Get_Line (it was changed in order to embrace the changes described in the phrase above).
SymbolTable.hs now shows the scope ID for each scope. This way each symbol can be traced to its corresponding scope. This was not effortless at all. The scope list is not ordered, so in order to maximize efficiency during the generation of the intermediate code we should not forget to sort the list according to it's scope ID. The scope IDs are in String format so we likely will apply a map with an "convert to Int" function, before ordering it.