Added Assignment Checking
Now the Symbol Table also includes a boolean value (Assigned type) for each entry (for each symbol), which represents whether or not that symbol/variable has been assigned a value properly. If True, then it has, if False then it has not. This allows us to check if a reference to a variable can be made. For example, only saying x is an Integer is not enough for us to include x in an expression such as x + 1. We need to assign x a value (an Integer value in this case) so that the expression can actually mean something that is defined. Thanks to this change, each entry of the Symbol Table has 3 values: the name, the type and the boolean which represents where the variable/symbol has been defined/assigned a value. If there's an error of this kind, there will be a "AssignErrorST" in the error list (third value of the result of this Symbol Table program).