Improved the new undefined variable checking
The previous version was still not complete. If a first assignment was made in a local scope, then once the scope was exited the table changes regarding the assignment would be nullified. One thing to note about this new feature is that there are at most two versions of the same variable in each Symbol Table: one for Assigned = True and one for Assigned = False (the latter always being the oldest one, the one the most on the right). During the Symbol Table construction if a first assignment is made inside a local scope, then it's children scopes will inherit it, and it's parent scope (where the declaration of the type was made) will also keep it. Naturally, before this local scope where the assignment was made there will be no knowledge of an assignment - only forever after the assignment was made will it be known.